chore: Update main.go to append LlamaGo server to llamaServers slice
The code changes in `main.go` modify the `llamaServers` slice by appending the string "llm_server_v2". This change is necessary to include the LlamaGo server in the list of servers to be processed. This commit message follows the convention of using the `chore` prefix to indicate a routine task or maintenance change.