chore: Update config.yaml to comment out the api_key field
The code changes in `config.yaml` comment out the `api_key` field by adding a '#' symbol at the beginning of the line. This change is made to prevent the API key from being exposed in the configuration file. This commit message follows the convention of using the `chore` prefix to indicate a routine task or maintenance change.