chore: Add new WebUi link to README.md
The code changes in `README.md` add a new WebUi link to the existing basic web UI section. This change allows users to access the new WebUi by visiting the specified URL. This commit message follows the convention of using the `chore` prefix to indicate a routine task or maintenance change.