chore: Update config.yaml to set the number of GPU layers to 30 and add tensor-split field
The code changes in `config.yaml` modify the `gpu-layers` field, setting it to "30" instead of "40". This change reduces the number of GPU layers used. Additionally, a new field `tensor-split` is added with the value "0.3,0.7" to specify the fraction of the model to offload to each GPU. This commit message follows the convention of using the `chore` prefix to indicate a routine task or maintenance change.